app/requirements.txt
streamlit
#psycopg2
psycopg2-binary
pandas
